Retrieve frame count.

+++ src/core/canvas.rs
@@ -2,7 +2,7 @@ extern crate gfx_hal as hal;
 
 pub trait Canvas
 {
-  fn image_count(&self) -> u32;
+  fn image_count(&self) -> usize;
   fn is_processing_frame(&self) -> bool;
   fn begin_frame(&mut self) -> Result<(), BeginFrameError>;
   fn end_frame(&mut self) -> Result<(), EndFrameError>;
